Explicitly treat free(NULL) as a no-op.
Previously, there was no explicit handling for this, so it would wind up treating the last few bytes of ROM as headers from the ORCA allocator (because the address calculation wrapped around). With actual IIGS ROMs, these bytes had values that ultimately led to it behaving as a no-op, but we should not rely on this.