Make #pragma rtl programs give the proper return value.
Previously, the return value was trashed, so the correct value was not returned. This affected #pragma rtl programs used as shell executables or potentially as a plugin/extension for another program, but it did not matter for PIFs or TIFs, since they do not have a return value.

Here is an example affected by this:

#pragma rtl
int main(void) {
        return 123;
}