realloc: When expanding, only copy up to size of old allocation.
Previously, realloc would always copy a number of bytes equal to the new size. This is needlessly slow if the new size is much larger than the old size. Depending on the sizes and the address of the old allocation, it can also occasionally cause soft switches or slot I/O locations to be accessed, which can have undesired effects.

This change fixes the problems by only copying up to the smaller of the two sizes (although the old size used reflects actual size allocated, which may be somewhat larger than what was requested). This fixes #1.